Revert "feat(ehr): photo-ID (driver's license) OCR + demographics/address AI suggestions (#8746)"
This reverts commit 165412476cd7b35235eb9629828d320f042c034a.
