+++ config/oystehr-core/zambdas.json
@@ -766,13 +766,6 @@
       "src": "src/subscriptions/document-reference/extract-insurance-card/index",
       "zip": ".dist/zips/extract-insurance-card.zip"
     },
-    "EXTRACT-PHOTO-ID": {
-      "name": "extract-photo-id",
-      "type": "subscription",
-      "runtime": "nodejs22.x",
-      "src": "src/subscriptions/document-reference/extract-photo-id/index",
-      "zip": ".dist/zips/extract-photo-id.zip"
-    },
     "ROTATE-INSURANCE-CARD-IMAGE": {
       "name": "rotate-insurance-card-image",
       "type": "http_auth",
@@ -2316,24 +2309,6 @@
         "channel": {
           "type": "rest-hook",
           "endpoint": "zapehr-lambda:#{ref/zambdas/EXTRACT-INSURANCE-CARD/id}"
-        },
-        "extension": [
-          {
-            "url": "http://zapehr.com/fhir/extension/SubscriptionTriggerEvent",
-            "valueString": "create"
-          }
-        ]
-      }
-    },
-    "EXTRACT_PHOTO_ID_SUBSCRIPTION": {
-      "resource": {
-        "resourceType": "Subscription",
-        "status": "active",
-        "reason": "Extract photo ID fields with OCR when an ID image is uploaded",
-        "criteria": "DocumentReference?type=http://loinc.org|55188-7&status=current",
-        "channel": {
-          "type": "rest-hook",
-          "endpoint": "zapehr-lambda:#{ref/zambdas/EXTRACT-PHOTO-ID/id}"
         },
         "extension": [
           {
